Started separation of concerns
By breaking up the uber-function into logical pieces

It makes it easier to test and understand.

If I don't have a browser I can for example replace window.location
